fix(set): handle system dialog box asynchronously for old API
When attempting to set the default system browser, a dialog box
is presented to the user.

Old behaviour:
The final print message would be output before the user had interacted
with the dialog box.

New behaviour:
The program waits until the user interacts with the dialog box and
outputs the correct message and at the correct time.